Pytorch36/Gated Convolution By Zhouge/requirements.txt
imageio==2.19.3
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
protobuf==3.19.4
scikit-image==0.19.2
scipy==1.8.1
tensorboard==2.9.1
tensorboardX==2.5.1
tensorflow-cpu==2.9.0
torch==1.10.1
torchaudio==0.10.1
torchsummary==1.5.1
torchvision==0.11.2
tqdm==4.64.0
